refactor: split hooks.ts into focused modules under hooks/
Move the 901-line hooks.ts into a hooks/ subdirectory with focused modules:

  extensions/hooks/types.ts       — HookResult, HookOutput interfaces
  extensions/hooks/matcher.ts     — matchesTool, normalizeToolInput
  extensions/hooks/runner.ts      — resolveShell, _resetShellCache, runHook, parseHookOutput
  extensions/hooks/events.ts      — getMatchingHooks, runEventHooks, runFireAndForgetHooks
  extensions/hooks/extension.ts   — createHooksExtension (the pi ExtensionFactory)
  extensions/hooks/index.ts       — barrel re-exporting all 5 modules
  extensions/hooks.ts             — 1-line: export * from './hooks/index.js'

hooks.test.ts stays at extensions/hooks.test.ts (unchanged import path).
factories.ts import of createHooksExtension unchanged.
No logic changes — pure reorganization. All 148 tests pass.
